# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Book Store",
        "description": "API para gestionar libros, autores, editoriales y categorias",
        "contact": {
            "email": "angelj.vazquez1@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/editorials": {
            "get": {
                "tags": [
                    "Editorials"
                ],
                "summary": "Listar editoriales",
                "description": "Devuelve una lista paginada de editoriales",
                "operationId": "getEditorials",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página para la paginación",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de editoriales paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Editorial"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/editorials?page=2"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Editorials"
                ],
                "summary": "Crear una nueva editorial",
                "description": "Crea una editorial con nombre, país y sitio web",
                "operationId": "storeEditorial",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "country",
                                    "website"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Editorial Alfa"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "España"
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "https://editorialalfa.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Editorial creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Editorial created successfully"
                                        },
                                        "editorial": {
                                            "$ref": "#/components/schemas/Editorial"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/editorials/{id}": {
            "get": {
                "tags": [
                    "Editorials"
                ],
                "summary": "Obtener una editorial por ID",
                "description": "Devuelve una editorial si el ID existe",
                "operationId": "getEditorialById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la editorial",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Editorial encontrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Editorial found successfully"
                                        },
                                        "editorial": {
                                            "$ref": "#/components/schemas/Editorial"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Editorial no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Editorial not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Editorials"
                ],
                "summary": "Actualizar una editorial",
                "description": "Actualiza los datos de una editorial existente por su ID",
                "operationId": "updateEditorial",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la editorial a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "country",
                                    "website"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Editorial Actualizada"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "México"
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "https://editorialactualizada.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Editorial actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Editorial updated successfully"
                                        },
                                        "editorial": {
                                            "$ref": "#/components/schemas/Editorial"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Editorial no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Editorial not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Editorials"
                ],
                "summary": "Eliminar una editorial",
                "description": "Elimina una editorial existente por su ID",
                "operationId": "deleteEditorial",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la editorial a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Editorial eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Editorial deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Editorial no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Editorial not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/editorials/by-name": {
            "post": {
                "tags": [
                    "Editorials"
                ],
                "summary": "Buscar editorial por nombre",
                "description": "Devuelve una editorial que coincida con el nombre proporcionado",
                "operationId": "getEditorialByName",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Alfaguara"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Editorial encontrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Editorial found successfully"
                                        },
                                        "editorial": {
                                            "$ref": "#/components/schemas/Editorial"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Editorial no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Editorial not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Editorial": {
                "title": "Editorial",
                "required": [
                    "name",
                    "country",
                    "website"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Planeta"
                    },
                    "country": {
                        "type": "string",
                        "example": "España"
                    },
                    "website": {
                        "type": "string",
                        "example": "https://www.editorialplaneta.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-02T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}